+++ src/networks/modules/Conv2d/codegen.cpp
@@ -24,7 +24,7 @@
 
 
 
-void conv2d_backward(float *inp, float size, float *out,
+void conv2d_backward(float *inp, int size, float *out,
                      float *dinp, float *dout,
                      std::string module_name, DT_tensor *node)
 {
